requirements.txt
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.6
django-appconf==1.0.6
django-heroku==0.3.1
django-imagekit==5.0.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pilkit==3.0
Pillow==10.1.0
psycopg2==2.9.9
razorpay==1.4.1
reportlab==4.0.9
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.6.0
